Added "allgames" page.
I have added a new page to allow the index page to only list categories, and the "allgames" page will handle the display of all the games and said categories. This looks cleaner as I want index page to be more of an intro to the website. I have also added a session_test.php page for debugging, found online.